+++ src/components/Schedule.astro
@@ -2,7 +2,7 @@
 import { getCollection, getEntry } from "astro:content";
 import { marked } from "marked";
 import DOMPurify from "isomorphic-dompurify";
-import { localizeTime, formatTime } from "../utils";
+import { localizeTime, formatTime } from "../utils/dateUtils";
 import { Badge } from "@astrojs/starlight/components";
 import { transformContentData } from "../utils/contentUtils";
