+++ src/content/docs/homeworks/homework 4.mdx
@@ -1,22 +1,28 @@
 ---
 title: 'Homework 4: SECURITY'
 description: Hacking a juice shop 🧃
-due: '2024-11-25T23:59:00'
-release: '2024-11-01T08:00:00'
+sidebar:
+  order: 34
+notes: >-
+  Watch [the kickoff video](https://youtu.be/5xExn6wqHnA) for help getting
+  started.
 relative_release:
-  week: 10
-  day: 0
+  week: 8
+  day: 1
   time: '08:00:00'
 relative_due:
-  week: 13
+  week: 9
   day: 0
-  time: '23:59:00'
+  time: '14:00:00'
 ---
 
-import { localizeTime, formatTime } from "../../../utils.ts";
 import { Content as Deadlines } from "../../../components/Deadlines.md";
+import { DueDate } from "../../../components/DueDate";
+import { Aside } from "@astrojs/starlight/components";
+import { marked } from "marked";
 
-**Due**: {formatTime(localizeTime(frontmatter.due))}
+<DueDate relativeDate={frontmatter.relative_due} />
+<DueDate relativeDate={frontmatter.relative_release} label="Release" />
 
 # {frontmatter.description}
